debundle/lowering: extract plan_references (+ PR #1633 bridge)
BAZEL_TEST_INVOCATIONS=none: file move.
